+++ src/rff2/vulkan/CPC2MapIterationStripe.hpp
@@ -4,7 +4,7 @@
 // Modified by GPT-5 on 2026-07-09, 2026-08-21, 2026-08-23
 // Modified by Opus 5 on 2026-08-05, 2026-08-07, 2026-08-13, 2026-08-15, 2026-08-17, 2026-08-18, 2026-08-19, 2026-08-24, 2026-08-25, 2026-08-26
 // Modified by Fable 5.1 on 2026-09-06
-// Modified by GPT-6 on 2026-09-08, 2026-09-11, 2026-09-15, 2026-09-16, 2026-09-20, 2026-09-22, 2026-09-23
+// Modified by GPT-6 on 2026-09-08, 2026-09-11, 2026-09-15, 2026-09-16, 2026-09-20, 2026-09-22, 2026-09-23, 2026-09-26
 //
 
 #pragma once
@@ -116,17 +116,17 @@ namespace merutilm::rff2 {
 
         void setInfo(double maxIteration, double normalMaxIteration, double zoomedMaxIteration) const;
 
-        void setTime(float currentSec, uint32_t frameIndex);
+        void setTime(double currentSec, uint32_t frameIndex);
 
         // Brings every animation phase up to this instant under the speeds in effect until now.
         // Call it before the timeline replaces any of them, or the new speed is charged for time
         // it was not running. A backwards or long jump is a seek, and re-derives them instead.
-        void advanceAnimationTo(float sec);
+        void advanceAnimationTo(double sec);
 
         void seekAnimationTo(float sec) { phases.seekTo(sec); }
 
         void setTimelinePhases(const std::array<double, ShaderAnimationPhases::TIMELINE_AXIS_COUNT> &values,
-                               float sec) { phases.setTimelinePhases(values, sec); }
+                               double sec) { phases.setTimelinePhases(values, sec); }
 
         [[nodiscard]] std::vector<uint32_t> specializationConstants() const override;
